refactor: move all toGrpc mappers into single file
