+++ src/main/kotlin/se/uulm/snowballr/backend/model/fetcher/FetcherInformation.kt
@@ -2,7 +2,6 @@ package se.uulm.snowballr.backend.model.fetcher
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import snowballr.Fetcher
 
 @Serializable
 data class FetcherInformation(
@@ -15,11 +14,3 @@ data class FetcherInformation(
     @SerialName("options_schema")
     val optionsSchema: Map<String, FetcherOptionsSchema>,
 )
-
-fun FetcherInformation.toGrpc(fetcherId: String): Fetcher.FetcherInformation = Fetcher.FetcherInformation.newBuilder()
-    .setId(fetcherId)
-    .setName(name)
-    .setDescription(description)
-    .addAllLinks(links.map { it.toGrpc() })
-    .putAllOptionsSchema(optionsSchema.mapValues { it.value.toGrpc() })
-    .build()
