+++ src/main/kotlin/se/uulm/snowballr/backend/model/fetcher/FetcherOptionsSchema.kt
@@ -2,7 +2,6 @@ package se.uulm.snowballr.backend.model.fetcher
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import snowballr.Fetcher
 
 @Serializable
 data class FetcherOptionsSchema(
@@ -17,11 +16,3 @@ data class FetcherOptionsSchema(
     @SerialName("default_value")
     val defaultValue: String?,
 )
-
-fun FetcherOptionsSchema.toGrpc(): Fetcher.FetcherOptionSchema = Fetcher.FetcherOptionSchema.newBuilder()
-    .setName(name)
-    .setDescription(description)
-    .setRequired(isRequired)
-    .setIsSecret(isSecret)
-    .apply { if (defaultValue != null) setDefaultValue(defaultValue) }
-    .build()
