+++ src/main/kotlin/se/uulm/snowballr/backend/model/fetcher/Link.kt
@@ -2,7 +2,6 @@ package se.uulm.snowballr.backend.model.fetcher
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import snowballr.Base
 
 @Serializable
 data class Link(
@@ -11,8 +10,3 @@ data class Link(
     @SerialName("url")
     val url: String,
 )
-
-fun Link.toGrpc(): Base.Link = Base.Link.newBuilder()
-    .setLabel(label)
-    .setUrl(url)
-    .build()
