+++ src/main/kotlin/se/uulm/snowballr/backend/model/outgoing/paper/FetcherPaperResponse.kt
@@ -3,10 +3,8 @@ package se.uulm.snowballr.backend.model.outgoing.paper
 import se.uulm.snowballr.backend.model.dto.paper.Author
 import se.uulm.snowballr.backend.model.dto.paper.Paper
 import se.uulm.snowballr.backend.model.dto.paper.PaperData
-import se.uulm.snowballr.backend.model.dto.paper.toGrpc
 import se.uulm.snowballr.backend.model.fetcher.FetcherMetadata
 import se.uulm.snowballr.backend.model.fetcher.FetcherPaper
-import snowballr.PaperOuterClass
 import java.util.UUID
 
 data class FetcherPaperResponse(
@@ -45,21 +43,3 @@ data class FetcherPaperResponse(
         )
     }
 }
-
-fun FetcherPaperResponse.toGrpc(): PaperOuterClass.Paper = PaperOuterClass.Paper.newBuilder()
-    .setId(id?.toString().orEmpty())
-    .setExternalId(externalId.orEmpty())
-    .setTitle(title)
-    .setAbstrakt(abstract)
-    .setYear(year)
-    .setPublisher(publisher)
-    .setPublicationName(publicationName)
-    .setPublicationType(publicationType)
-    .addAllAuthors(authors.toGrpc())
-    .putAllFetcherMetadata(fetcherMetadata)
-    .addAllBackwardReferencedIds(backwardReferencedIds.map { it.toString() })
-    .build()
-
-fun List<FetcherPaperResponse>.toGrpc(): PaperOuterClass.Paper.List = PaperOuterClass.Paper.List.newBuilder()
-    .addAllPapers(this.map { it.toGrpc() })
-    .build()
