+++ src/main/kotlin/se/uulm/snowballr/backend/model/outgoing/project/ProjectDecisionStatistics.kt
@@ -1,12 +1,5 @@
 package se.uulm.snowballr.backend.model.outgoing.project
 
-import snowballr.ProjectOuterClass
-
 data class ProjectDecisionStatistics(
     val statistics: List<ProjectDecisionCount>,
 )
-
-fun ProjectDecisionStatistics.toGrpc(): ProjectOuterClass.Project.Information.DecisionStatistics =
-    ProjectOuterClass.Project.Information.DecisionStatistics.newBuilder()
-        .addAllStatistics(statistics.map { it.toGrpc() })
-        .build()
